requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
readability-lxml==0.8.1
sentence-transformers>=2.2.2
nltk==3.8.1
faiss-cpu>=1.9.0.post1
pydantic>=2.5.0
python-multipart>=0.0.6
requests>=2.31.0
urllib3>=2.1.0
lxml[html_clean]>=4.9.3
numpy>=1.26.0
scikit-learn>=1.3.2
